Move export functionality into reporting module to avoid patching
