+++ modules/reporting/app/views/cost_reports/index.html.erb
@@ -43,6 +43,16 @@ See COPYRIGHT and LICENSE files for more details.
       </h2>
     </div>
     <ul class="toolbar-items">
+      <% if User.current.allowed_to? :export_work_packages, @project, global: @project.nil? %>
+        <li class="toolbar-item">
+          <%= link_to(t(:export_to_excel),
+                      { controller: "cost_reports" ,
+                        action: :index,
+                        format: 'xls',
+                        project_id: @project },
+                      class: "button icon-export-xls-descr") %>
+        </li>
+      <% end %>
       <%= call_hook(:view_cost_report_toolbar) %>
     </ul>
   </div>
