--- modules/xls_export/app/models/xls_export/cost_reports/export.rb
@@ -1,14 +1,0 @@
-module XlsExport
-  module CostReports
-    class Export < ::Export
-      acts_as_attachable view_permission: :export_work_packages,
-                         add_permission: :export_work_packages,
-                         delete_permission: :export_work_packages,
-                         only_user_allowed: true
-
-      def ready?
-        attachments.any?
-      end
-    end
-  end
-end
