--- modules/xls_export/app/views/hooks/xls_report/_view_cost_report_toolbar.html.erb
@@ -1,10 +1,0 @@
-<% if User.current.allowed_to? :export_work_packages, @project, global: @project.nil? %>
-  <li class="toolbar-item">
-    <%= link_to(t(:export_to_excel),
-                { controller: "cost_reports" ,
-                  action: :index,
-                  format: 'xls',
-                  project_id: @project },
-                class: "button icon-export-xls-descr") %>
-  </li>
-<% end %>
