--- modules/xls_export/lib/open_project/xls_export/hooks/cost_report_hook.rb
@@ -1,5 +1,0 @@
-module OpenProject::XlsExport::Hooks
-  class CostReportHook < OpenProject::Hook::ViewListener
-    render_on :view_cost_report_toolbar, partial: 'hooks/xls_report/view_cost_report_toolbar'
-  end
-end
